+++ ncs/propagators/alldifferent_lopez_ortiz.py
@@ -1,6 +1,5 @@
 from typing import Optional
 
-import numba
 import numpy as np
 from numba import jit
 from numpy.typing import NDArray
@@ -143,12 +142,12 @@ def path_max(t: NDArray, i: int) -> int:
 
 @jit(nopython=True, nogil=True)
 def compute_domains(size: int, domains: NDArray) -> Optional[NDArray]:
-    rank_domains = np.hstack((domains, np.zeros((size, 2), dtype=numba.int32)))
+    rank_domains = np.hstack((domains, np.zeros((size, 2), dtype=np.uint16)))
     bounds_nb = 2 * size + 2
-    bounds = np.zeros(bounds_nb, dtype=numba.int32)
-    t = np.zeros(bounds_nb, dtype=numba.int32)
-    d = np.zeros(bounds_nb, dtype=numba.int32)
-    h = np.zeros(bounds_nb, dtype=numba.int32)
+    bounds = np.zeros(bounds_nb, dtype=np.int32)
+    t = np.zeros(bounds_nb, dtype=np.int32)
+    d = np.zeros(bounds_nb, dtype=np.int32)
+    h = np.zeros(bounds_nb, dtype=np.int32)
     min_sorted_vars = np.argsort(rank_domains[:, MIN])
     max_sorted_vars = np.argsort(rank_domains[:, MAX])
     nb = compute_nb(size, rank_domains, min_sorted_vars, max_sorted_vars, bounds)
